+++ frontend/src/service/index.ts
@@ -1,0 +1,17 @@
+import * as chatService from './chatService';
+import * as defenceService from './defenceService';
+import * as documentService from './documentService';
+import * as emailService from './emailService';
+import * as healthService from './healthService';
+import * as levelService from './levelService';
+import * as systemRoleService from './systemRoleService';
+
+export {
+	chatService,
+	defenceService,
+	documentService,
+	emailService,
+	healthService,
+	levelService,
+	systemRoleService,
+};
